FIX fixed all the C4244 warnings

+++ SSPSolution/SSPSolution/Animation.cpp
@@ -32,7 +32,7 @@ Animation::Animation()
 	float endFrame = animatedJointsList->keyframes[animatedJointsList->keyframeCount - 1].timeValue;
 
 	/*Initialize the stack with a default "IDLE" animation.*/
-	Push(currentAnimation, true, startFrame, endFrame, 0);
+	Push(currentAnimation, true, (int)startFrame, (int)endFrame, 0);
 }
 
 Animation::~Animation()
@@ -119,7 +119,7 @@ void Animation::Interpolate(float currentTime, std::vector<XMFLOAT4X4> interpola
 		/*The current time is the first keyframe.*/
 		if (currentTime <= animationStack.top().startFrame)
 		{
-			int startFrame = animationStack.top().startFrame;
+			int startFrame = (int)animationStack.top().startFrame;
 
 			XMFLOAT3 tempTrans(animatedJoint.keyframes[startFrame].translation);
 			XMFLOAT3 tempScale(animatedJoint.keyframes[startFrame].scale);
@@ -136,7 +136,7 @@ void Animation::Interpolate(float currentTime, std::vector<XMFLOAT4X4> interpola
 		/*The current time is the last keyframe.*/
 		else if (currentTime >= animationStack.top().endFrame)
 		{
-			int endFrame = animationStack.top().endFrame;
+			int endFrame = (int)animationStack.top().endFrame;
 
 			XMFLOAT3 tempTrans(animatedJoint.keyframes[endFrame].translation);
 			XMFLOAT3 tempScale(animatedJoint.keyframes[endFrame].scale);
@@ -252,8 +252,8 @@ void Animation::Push(int currentAnimation, bool isLooping, int startFrame, int e
 
 	AnimationClip clip;
 	clip.isLooping = isLooping;
-	clip.startFrame = startFrame;
-	clip.endFrame = endFrame;
+	clip.startFrame = (float)startFrame;
+	clip.endFrame = (float)endFrame;
 
 	animationStack.push(clip);
 }
